requirements.txt
google-api-python-client>=2.79.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
requests>=2.28.1
python-dotenv>=1.0.0
loguru>=0.6.0